+++ tse_analytics/toolbox/data_plot/data_plot_widget.py
@@ -161,7 +161,7 @@ class DataPlotWidget(QWidget):
 
         df = self._get_timeline_plot_df(selected_variables_dict)
 
-        (
+        plot = (
             so
             .Plot(
                 df,
@@ -169,8 +169,14 @@ class DataPlotWidget(QWidget):
                 color=by,
             )
             .pair(y=list(selected_variables_dict))
-            .add(so.Line(linewidth=self.linewidth_spin_box.value()), so.Agg(func="mean"))  # Line with mean estimate
-            .add(so.Band(alpha=0.15), so.Est(errorbar=error_bar))  # Shaded CI band
+            .add(so.Line(linewidth=self.linewidth_spin_box.value()), so.Agg(func="mean"))
+        )
+
+        if error_bar is not None:
+            plot = plot.add(so.Band(alpha=0.15), so.Est(errorbar=error_bar))
+
+        (
+            plot
             .scale(
                 color=so.Nominal(palette, order=df[by].cat.categories.tolist()) if by else palette,
             )